restructure library: add color_detect.py, ir_light_bulb.py files
